refactor: create rehype generator
